pyproject.toml
[project]
name = "websweeper"
version = "0.1.0"
description = "Config-driven Playwright automation framework"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.40",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
websweeper = "websweeper.cli:cli"

[project.entry-points."websweeper.extensions"]
finance = "extensions.finance.actions:finance_group"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/websweeper", "extensions"]

[tool.hatch.build.targets.editable]
packages = ["src/websweeper", "extensions"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
